+++ src/main/java/com/bencompany/expensetracker/model/Payment.java
@@ -17,7 +17,7 @@ public class Payment {
     @JoinColumn(name = "user_id", nullable = false) // creates the FK column in Payment table
     private User user;
 
-    private int amount;
+    private double amount;
     private String category;
     private LocalDateTime timeCreated;
     private String notes;
@@ -29,14 +29,14 @@ public class Payment {
     //getters
     public UUID getPaymentId(){return paymentID;}
     public User getUser(){return user;}
-    public int getAmount(){return amount;}
+    public double getAmount(){return amount;}
     public String getCategory(){return category;}
     public LocalDateTime getTimeCreated(){return timeCreated;}
     public String getNotes(){return notes;}
 
     //setters
     public void setUser(User user) { this.user = user; }
-    public void setAmount(int amount) { this.amount = amount; }
+    public void setAmount(double amount) { this.amount = amount; }
     public void setCategory(String category) { this.category = category; }
     public void setNotes(String notes) { this.notes = notes; }
